+++ lib/brontosaurus_flutter.dart
@@ -3,5 +3,6 @@ library brontosaurus_flutter;
 export "src/config.dart";
 export "src/declare.dart";
 export "src/lite.dart";
+export "src/parse.dart";
 export "src/token.dart";
 export "src/view.dart";